requirements.txt
fastapi
uvicorn[standard]
pydantic
icmplib
dnspython
qasync
